requirements.txt
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
insightface>=0.7.3
onnxruntime>=1.16.0
opencv-python>=4.8.0
numpy<2.0.0
scipy>=1.11.0
mediapipe>=0.10.0,<0.11
protobuf<5